+++ community/boilerplates/graphql-servers/nodejs-apollo/README.md
@@ -1,3 +1,3 @@
 # (MOVED) GraphQL server using NodeJS and Apollo
 
-**This folder has been moved to [remote-schemas/zeit-now/nodejs/apollo-sequelize](../../remote-schemas/zeit-now/nodejs/apollo-sequelize)**
+**This folder has been moved to [remote-schemas/aws-lambda/nodejs](../../remote-schemas/aws-lambda/nodejs)**
