+++ community/boilerplates/graphql-servers/nodejs-express/README.md
@@ -1,3 +1,3 @@
 # (MOVED) GraphQL server using NodeJS-Express
 
-**This folder has been moved to [remote-schemas/zeit-now/nodejs/express-graphqljs-sequelize](../../remote-schemas/zeit-now/nodejs/express-graphqljs-sequelize)**
+**This folder has been moved to [remote-schemas/aws-lambda/nodejs](../../remote-schemas/aws-lambda/nodejs)**
