+++ community/boilerplates/remote-schemas/aws-lambda/nodejs/.gitignore
@@ -1,0 +1,2 @@
+node_modules/
+*.zip
