--- community/boilerplates/remote-schemas/aws-lambda/nodejs/apollo-sequelize/.gitignore
@@ -1,3 +1,0 @@
-node_modules
-build
-*.zip
