+++ community/boilerplates/remote-schemas/aws-lambda/nodejs/package.json
@@ -2,7 +2,7 @@
   "name": "aws-lambda-nodejs",
   "version": "1.0.0",
   "description": "",
-  "main": "graphql.js",
+  "main": "lambdaCtx.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -10,11 +10,8 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "apollo-server": "^2.2.2",
     "apollo-server-lambda": "^2.1.0",
     "graphql": "^0.13.1",
-    "pg": "^7.6.0",
-    "pg-hstore": "^2.3.2",
-    "sequelize": "^4.41.0"
+    "graphql-tag": "^2.10.1"
   }
 }
